fix: error de unas vainas
