+++ src/main/java/org/tvl/tvlooker/domain/strategy/recommendation/MatrixFactorizationStrategy.java
@@ -1,7 +1,10 @@
 package org.tvl.tvlooker.domain.strategy.recommendation;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.tvl.tvlooker.domain.data_structure.ScoredItem;
 import org.tvl.tvlooker.domain.data_structure.SVDFactors;
+import org.tvl.tvlooker.domain.exception.NoDataProviderException;
 import org.tvl.tvlooker.domain.model.dto.Item;
 import org.tvl.tvlooker.domain.model.dto.User;
 import org.tvl.tvlooker.domain.motor.utils.RecommendationContext;
@@ -21,13 +24,20 @@ import java.util.stream.Collectors;
  */
 public class MatrixFactorizationStrategy implements RecommendationStrategy {
 
+    private static final Logger logger = LoggerFactory.getLogger(MatrixFactorizationStrategy.class);
     private static final double MAX_RATING = 5.0;
     private static final String EXPLANATION = "Based on your overall preferences";
     private static final String PROVIDER_ID = "svd-factors";
 
     @Override
     public List<ScoredItem> recommend(User user, List<Item> candidateItems, RecommendationContext context) {
-        SVDFactors factors = context.getData(PROVIDER_ID, SVDFactors.class);
+        SVDFactors factors;
+        try {
+            factors = context.getData(PROVIDER_ID, SVDFactors.class);
+        } catch (NoDataProviderException e) {
+            logger.debug("No SVD factors provider available, returning empty recommendations");
+            return List.of();
+        }
 
         if (factors == null || factors.getLatentFactors() == 0) {
             return List.of();
